Some new features some bug fixes
+ update to okhttpv3 (this time for real)

in okhttp v2 every new okhttp client
would share one connection pool.
In Okhttpv3 they changed it so that
every new okhttp client has it's own
connection pool and every
connection pool can hold open connections
for up to 5 minutes due to this it's easy
to run out of open file descriptors on
android given our previous code design
in which we would create a new okhttp
client for every connection. We now have only
one global okhttp client  that we use for all
connections, this client is then sometimes customized
with custom connection/read/write timeout settings.
Anyway the behaviour should be the same for users
and we benefit from new changes in okhttp so it's
a win-win.

+ add Local Search for db
+ clear cache now also deletes empty directories in MiManga directory
+ Some behaviour changes for the searching for updates notification
1. clicking on the notification will now get you back to the Main Screen (MainActivity) instead of cancelling the update
2. cancelling the update process now works via the cancel action
3. should now be cancellable even when new chapters have been found